refactor(mitm-addon): split websocket usage tests by lifecycle
