--- crates/runner/mitm-addon/tests/test_model_provider_websocket_usage.py
@@ -1,1566 +1,0 @@
-"""Tests for model-provider WebSocket usage reporting paths."""
-
-import json
-import uuid
-from pathlib import Path
-from typing import Literal
-
-import pytest
-from mitmproxy import http
-
-import flow_metadata_keys as metadata_keys
-import mitm_addon
-import response_streaming
-import usage
-from tests.jsonl_log_helpers import jsonl_exists_after_flush, read_jsonl_entries_after_flush
-from tests.model_provider_flow_helpers import (
-    make_openai_responses_websocket_flow,
-    model_provider_usage_sources,
-    model_usage_source_entries,
-)
-from tests.model_provider_websocket_helpers import (
-    ScheduledWebSocketTrim,
-    capture_deferred_websocket_trims,
-    feed_websocket_client_message,
-    feed_websocket_server_message,
-    feed_websocket_server_text_message,
-    openai_websocket_usage_frame,
-    set_websocket_message,
-)
-from tests.usage_helpers import (
-    compact_observation_quantities,
-    compact_observation_rows,
-)
-
-
-@pytest.fixture(autouse=True)
-def deferred_websocket_trim_scheduler(
-    monkeypatch: pytest.MonkeyPatch,
-) -> list[ScheduledWebSocketTrim]:
-    return capture_deferred_websocket_trims(monkeypatch)
-
-
-def _assert_usage_event_rows(
-    events: list[dict],
-    resource_field: Literal["provider", "model"],
-    expected_rows: list[tuple[str, str, int]],
-) -> None:
-    actual_rows = (
-        compact_observation_rows(events)
-        if resource_field == "model"
-        else [(event[resource_field], event["category"], event["quantity"]) for event in events]
-    )
-    assert len(actual_rows) == len(expected_rows)
-    assert sorted(actual_rows) == sorted(expected_rows)
-
-    idempotency_keys = [event["idempotencyKey"] for event in events]
-    assert len(set(idempotency_keys)) == len(idempotency_keys)
-    for key in idempotency_keys:
-        uuid.UUID(key)
-
-
-def _openai_websocket_zero_usage_frame(response_id: str, *, model: str | None = "gpt-5.5") -> bytes:
-    response: dict = {
-        "id": response_id,
-        "usage": {"input_tokens": 0, "output_tokens": 0},
-    }
-    if model is not None:
-        response["model"] = model
-    return json.dumps(
-        {
-            "type": "response.completed",
-            "response": response,
-        }
-    ).encode()
-
-
-def _openai_websocket_created_frame(response_id: str) -> bytes:
-    return json.dumps(
-        {
-            "type": "response.created",
-            "response": {"id": response_id},
-        }
-    ).encode()
-
-
-class TestModelProviderWebSocketPrewarmUsage:
-    """Tests for exact non-generating response source exclusion."""
-
-    @pytest.fixture(autouse=True)
-    def _sync_usage_delivery(self, sync_usage_executor, usage_webhook_api):
-        self._usage_webhook_api = usage_webhook_api
-
-    def test_model_websocket_ignores_bound_prewarm_and_reports_normal_input_only_turn(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        sensitive_marker = "prewarm-sensitive-marker"
-        prewarm_request = json.dumps(
-            {
-                "type": "response.create",
-                "input": [{"role": "user", "content": sensitive_marker * 300}],
-                "tools": [{"name": "test-tool", "description": "x" * 5000}],
-                "generate": False,
-            }
-        ).encode()
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_client_message(flow, prewarm_request)
-            feed_websocket_server_message(flow, _openai_websocket_created_frame("warm-1"))
-            prewarm_usage = openai_websocket_usage_frame(
-                "warm-1",
-                input_tokens=6050,
-                output_tokens=0,
-            )
-            feed_websocket_server_message(flow, prewarm_usage)
-
-            feed_websocket_client_message(
-                flow,
-                json.dumps({"type": "response.create", "input": []}).encode(),
-            )
-            feed_websocket_server_message(flow, _openai_websocket_created_frame("resp-1"))
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "resp-1",
-                    input_tokens=10,
-                    output_tokens=0,
-                ),
-            )
-            feed_websocket_server_message(flow, prewarm_usage)
-            mitm_addon.websocket_end(flow)
-            usage.flush_usage_events(trigger="test")
-
-        expected_rows = [("gpt-5.5", "tokens.input", 10)]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            expected_rows,
-        )
-        ignored_entries = [
-            entry
-            for entry in model_usage_source_entries(flow)
-            if entry.get("disposition") == "ignored"
-        ]
-        [ignored_entry] = ignored_entries
-        assert ignored_entry["reason"] == "responses_generate_false"
-        assert ignored_entry["provider_response_id"] == "warm-1"
-        assert ignored_entry["source_id"] == f"{flow.id}:warm-1"
-        assert ignored_entry["usage"] == {"tokens.input": 6050}
-        assert ignored_entry["usage_events"] == []
-        assert ignored_entry["model_usage_observations"] == []
-        assert ignored_entry["url"] == "https://api.openai.com/v1/responses"
-        proxy_log = Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-        assert sensitive_marker not in proxy_log.read_text()
-        assert model_provider_usage_sources(flow) == {}
-        assert response_streaming.is_model_websocket_usage_enabled(flow) is False
-        assert "_model_websocket_prewarm_state" not in flow.metadata
-
-    def test_model_websocket_unbound_prewarm_usage_fails_open(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_client_message(
-                flow,
-                json.dumps({"type": "response.create", "generate": False}).encode(),
-            )
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "warm-without-created",
-                    input_tokens=12,
-                    output_tokens=0,
-                ),
-            )
-            feed_websocket_server_message(
-                flow,
-                _openai_websocket_created_frame("response-after-unbound-terminal"),
-            )
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "response-after-unbound-terminal",
-                    input_tokens=4,
-                    output_tokens=0,
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 12),
-            ("gpt-5.5", "tokens.input", 4),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            expected_rows,
-        )
-        assert not any(
-            entry.get("disposition") == "ignored" for entry in model_usage_source_entries(flow)
-        )
-
-    def test_model_websocket_prewarm_with_missing_created_id_fails_open(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_client_message(
-                flow,
-                json.dumps({"type": "response.create", "generate": False}).encode(),
-            )
-            feed_websocket_server_message(
-                flow,
-                json.dumps({"type": "response.created", "response": {}}).encode(),
-            )
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "unbound-response",
-                    input_tokens=6,
-                    output_tokens=0,
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        expected_rows = [("gpt-5.5", "tokens.input", 6)]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            expected_rows,
-        )
-
-    @pytest.mark.parametrize(
-        "client_request",
-        [
-            b'{"type":"response.create"}',
-            b'{"type":"response.create","generate":true}',
-            b'{"type":"response.create","generate":"false"}',
-            b'{"type":"response.create","generate":true,"generate":false}',
-            b'{"type":"other","type":"response.create","generate":false}',
-        ],
-    )
-    def test_model_websocket_non_prewarm_requests_retain_input_only_usage(
-        self,
-        tmp_path,
-        real_flow,
-        client_request: bytes,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_client_message(flow, client_request)
-            feed_websocket_server_message(flow, _openai_websocket_created_frame("resp-input-only"))
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "resp-input-only",
-                    input_tokens=9,
-                    output_tokens=0,
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        expected_rows = [("gpt-5.5", "tokens.input", 9)]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            expected_rows,
-        )
-
-    def test_model_websocket_conflicting_created_ids_fail_open(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_client_message(
-                flow,
-                json.dumps({"type": "response.create", "generate": False}).encode(),
-            )
-            feed_websocket_server_message(
-                flow,
-                b'{"type":"response.created","response":{"id":"first","id":"second"}}',
-            )
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "second",
-                    input_tokens=11,
-                    output_tokens=0,
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        expected_rows = [("gpt-5.5", "tokens.input", 11)]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            expected_rows,
-        )
-
-    def test_model_websocket_conflicting_terminal_ids_fail_open(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_client_message(
-                flow,
-                json.dumps({"type": "response.create", "generate": False}).encode(),
-            )
-            feed_websocket_server_message(
-                flow,
-                _openai_websocket_created_frame("warm-ambiguous"),
-            )
-            feed_websocket_server_message(
-                flow,
-                b'{"type":"response.completed","response":'
-                b'{"id":"other","id":"warm-ambiguous","model":"gpt-5.5",'
-                b'"usage":{"input_tokens":13,"output_tokens":0}}}',
-            )
-            usage.flush_usage_events(trigger="test")
-
-        expected_rows = [("gpt-5.5", "tokens.input", 13)]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            expected_rows,
-        )
-        assert not any(
-            entry.get("disposition") == "ignored" for entry in model_usage_source_entries(flow)
-        )
-
-    def test_model_websocket_malformed_client_frame_retires_unbound_prewarm(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_client_message(
-                flow,
-                json.dumps({"type": "response.create", "generate": False}).encode(),
-            )
-            feed_websocket_client_message(flow, b'{"type":"response.create"')
-            feed_websocket_server_message(flow, _openai_websocket_created_frame("resp-after-bad"))
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "resp-after-bad",
-                    input_tokens=8,
-                    output_tokens=0,
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        expected_rows = [("gpt-5.5", "tokens.input", 8)]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            expected_rows,
-        )
-
-    def test_model_websocket_prewarm_state_isolated_by_flow(self, tmp_path, real_flow):
-        prewarm_flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        normal_flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(prewarm_flow)
-        mitm_addon.responseheaders(normal_flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_client_message(
-                prewarm_flow,
-                json.dumps({"type": "response.create", "generate": False}).encode(),
-            )
-            feed_websocket_server_message(
-                prewarm_flow,
-                _openai_websocket_created_frame("shared-response-id"),
-            )
-            feed_websocket_server_message(
-                prewarm_flow,
-                openai_websocket_usage_frame(
-                    "shared-response-id",
-                    input_tokens=5,
-                    output_tokens=0,
-                ),
-            )
-
-            feed_websocket_client_message(
-                normal_flow,
-                json.dumps({"type": "response.create"}).encode(),
-            )
-            feed_websocket_server_message(
-                normal_flow,
-                _openai_websocket_created_frame("shared-response-id"),
-            )
-            feed_websocket_server_message(
-                normal_flow,
-                openai_websocket_usage_frame(
-                    "shared-response-id",
-                    input_tokens=7,
-                    output_tokens=0,
-                ),
-            )
-            feed_websocket_client_message(
-                normal_flow,
-                json.dumps({"type": "response.create", "generate": False}).encode(),
-            )
-            feed_websocket_server_message(
-                normal_flow,
-                _openai_websocket_created_frame("second-prewarm"),
-            )
-            feed_websocket_server_message(
-                normal_flow,
-                openai_websocket_usage_frame(
-                    "second-prewarm",
-                    input_tokens=6,
-                    output_tokens=0,
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        expected_rows = [("gpt-5.5", "tokens.input", 7)]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            expected_rows,
-        )
-        ignored_entries = [
-            entry
-            for entry in model_usage_source_entries(prewarm_flow)
-            if entry.get("disposition") == "ignored"
-        ]
-        assert {
-            (entry["flow_id"], entry["provider_response_id"], entry["usage"]["tokens.input"])
-            for entry in ignored_entries
-        } == {
-            (prewarm_flow.id, "shared-response-id", 5),
-            (normal_flow.id, "second-prewarm", 6),
-        }
-
-
-class TestModelProviderWebSocketUsageSourceRelease:
-    """Tests for sources that cannot be delivered to the usage webhook."""
-
-    def test_model_websocket_missing_context_releases_positive_source(
-        self, tmp_path, real_flow, mitm_ctx
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        flow.metadata[metadata_keys.VM_SANDBOX_AUTH_KEY] = ""
-        proxy_log = Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-
-        with mitm_ctx(api_url="https://api.vm0.ai"):
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "resp_ws_missing_context",
-                    input_tokens=10,
-                    output_tokens=4,
-                ),
-            )
-
-        assert model_provider_usage_sources(flow) == {}
-        entries = read_jsonl_entries_after_flush(proxy_log)
-        [entry] = [entry for entry in entries if entry.get("type") == "usage_underbilling"]
-        [observation_entry] = [
-            entry for entry in entries if entry.get("type") == "model_usage_observation"
-        ]
-        assert not any(entry.get("type") == "model_usage_source" for entry in entries)
-        assert entry["type"] == "usage_underbilling"
-        assert entry["reason"] == "missing_reporting_context"
-        assert entry["underbilling_class"] == "confirmed"
-        assert entry["run_id"] == "run-abc-123"
-        assert entry["firewall_name"] == "model-provider:openai-api-key"
-        assert entry["missing_sandbox_token"] is True
-        assert entry["missing_api_url"] is False
-        assert observation_entry["level"] == "warn"
-        assert (
-            observation_entry["message"]
-            == "Cannot report model usage observation: missing sandbox_token or api_url"
-        )
-
-    def test_model_websocket_missing_api_url_releases_positive_source(
-        self, tmp_path, real_flow, mitm_ctx
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        proxy_log = Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-
-        with mitm_ctx(api_url=""):
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "resp_ws_missing_api_url",
-                    input_tokens=10,
-                    output_tokens=4,
-                ),
-            )
-
-        assert model_provider_usage_sources(flow) == {}
-        [entry] = [
-            entry
-            for entry in read_jsonl_entries_after_flush(proxy_log)
-            if entry.get("type") == "usage_underbilling"
-        ]
-        assert entry["reason"] == "missing_reporting_context"
-        assert entry["missing_sandbox_token"] is False
-        assert entry["missing_api_url"] is True
-
-    def test_model_websocket_missing_context_releases_zero_only_source(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        flow.metadata[metadata_keys.VM_SANDBOX_AUTH_KEY] = ""
-        proxy_log = Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-
-        feed_websocket_server_message(
-            flow,
-            openai_websocket_usage_frame(
-                "resp_ws_zero_missing_context",
-                input_tokens=0,
-                output_tokens=0,
-            ),
-        )
-
-        assert model_provider_usage_sources(flow) == {}
-        assert not jsonl_exists_after_flush(proxy_log)
-
-    def test_model_websocket_missing_api_url_releases_zero_only_source(
-        self, tmp_path, real_flow, mitm_ctx
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        proxy_log = Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-
-        with mitm_ctx(api_url=""):
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "resp_ws_zero_missing_api_url",
-                    input_tokens=0,
-                    output_tokens=0,
-                ),
-            )
-
-        assert model_provider_usage_sources(flow) == {}
-        assert not jsonl_exists_after_flush(proxy_log)
-
-
-class TestModelProviderWebSocketUsage:
-    """Tests for model-provider WebSocket usage reporting."""
-
-    @pytest.fixture(autouse=True)
-    def _sync_usage_delivery(self, sync_usage_executor, usage_webhook_api):
-        self._usage_webhook_api = usage_webhook_api
-
-    def _run_websocket_messages_and_end(self, flow: http.HTTPFlow, *messages: bytes):
-        with self._usage_webhook_api() as webhook:
-            for message in messages:
-                feed_websocket_server_message(flow, message)
-            mitm_addon.websocket_end(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def _run_websocket_message_and_end(self, flow: http.HTTPFlow):
-        with self._usage_webhook_api() as webhook:
-            mitm_addon.websocket_message(flow)
-            mitm_addon.websocket_end(flow)
-            usage.flush_usage_events(trigger="test")
-        return webhook
-
-    def test_full_pipeline_model_websocket_reports_usage(self, tmp_path, real_flow):
-        """Codex Responses WebSocket frames should bill like SSE events."""
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        assert flow.metadata["model_websocket_usage_enabled"] is True
-        assert "model_json_usage_finish" not in flow.metadata
-        assert "model_sse_usage_finish" not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER not in flow.metadata
-        assert metadata_keys.STREAM_BUFFER_STATE not in flow.metadata
-
-        set_websocket_message(
-            flow,
-            from_client=False,
-            content=json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "id": "resp_ws_1",
-                        "model": "gpt-5.5",
-                        "usage": {
-                            "input_tokens": 50,
-                            "output_tokens": 20,
-                            "input_tokens_details": {
-                                "cached_tokens": 10,
-                                "cache_write_tokens": 15,
-                            },
-                        },
-                    },
-                }
-            ).encode(),
-        )
-
-        webhook = self._run_websocket_message_and_end(flow)
-
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-        assert model_provider_usage_sources(flow) == {}
-        _assert_usage_event_rows(
-            webhook.usage_events(),
-            "provider",
-            [
-                ("gpt-5.5", "tokens.input", 25),
-                ("gpt-5.5", "tokens.output", 20),
-                ("gpt-5.5", "tokens.cache_read", 10),
-                ("gpt-5.5", "tokens.cache_creation", 15),
-            ],
-        )
-        [source_entry] = model_usage_source_entries(flow)
-        assert source_entry["flow_id"] == flow.id
-        assert source_entry["source_id"] == f"{flow.id}:resp_ws_1"
-        assert source_entry["provider_response_id"] == "resp_ws_1"
-        assert source_entry["transport"] == "websocket"
-        assert source_entry["buffer_mode"] == "source"
-        assert source_entry["method"] == "GET"
-        assert source_entry["url"] == "https://api.openai.com/v1/responses"
-        assert all(event["buffer_accepted"] is True for event in source_entry["usage_events"])
-        assert {event["source_idempotency_key"] for event in source_entry["usage_events"]} == {
-            event["idempotencyKey"] for event in webhook.usage_events()
-        }
-        assert all(
-            observation["buffer_accepted"] is True
-            for observation in source_entry["model_usage_observations"]
-        )
-        assert {
-            observation["source_idempotency_key"]
-            for observation in source_entry["model_usage_observations"]
-        } == {
-            observation["idempotencyKey"]
-            for observation in webhook.model_usage_observation_events()
-        }
-
-    def test_model_websocket_work_limit_warns_and_later_frame_reports(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        proxy_log = Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-        over_budget_frame = (
-            b'{"type":"response.completed","response":{"id":"resp_partial",'
-            b'"model":"gpt-5.5","usage":{"input_tokens":100,"output_tokens":40}},'
-            b'"padding":[' + b",".join([b"0"] * 40_000) + b"]}"
-        )
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_server_message(flow, over_budget_frame)
-
-            assert flow.websocket is not None
-            assert flow.websocket.messages[-1].content == over_budget_frame
-            assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-            assert model_provider_usage_sources(flow) == {}
-
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "resp_after_limit",
-                    input_tokens=7,
-                    output_tokens=3,
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        warning_entries = [
-            entry
-            for entry in read_jsonl_entries_after_flush(proxy_log)
-            if entry.get("message") == "Model provider WebSocket usage extraction failed"
-        ]
-        [warning] = warning_entries
-        assert warning["level"] == "warn"
-        assert warning["type"] == "usage_event"
-        assert warning["usage_protocol"] == "openai_responses_websocket"
-        assert warning["error"] == "work_limit_exceeded"
-        assert model_provider_usage_sources(flow) == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 7),
-            ("gpt-5.5", "tokens.output", 3),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-
-    def test_full_pipeline_model_websocket_reports_multiple_response_ids(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            openai_websocket_usage_frame(
-                "resp_ws_1",
-                input_tokens=10,
-                output_tokens=4,
-            ),
-            openai_websocket_usage_frame(
-                "resp_ws_2",
-                input_tokens=3,
-                output_tokens=2,
-            ),
-        )
-
-        assert model_provider_usage_sources(flow) == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 10),
-            ("gpt-5.5", "tokens.output", 4),
-            ("gpt-5.5", "tokens.input", 3),
-            ("gpt-5.5", "tokens.output", 2),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-
-    def test_model_websocket_late_same_id_frame_after_release_is_duplicate(
-        self, tmp_path, real_flow
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            openai_websocket_usage_frame(
-                "resp_ws_1",
-                input_tokens=20,
-                output_tokens=12,
-            ),
-            openai_websocket_usage_frame(
-                "resp_ws_1",
-                input_tokens=10,
-                output_tokens=7,
-            ),
-        )
-
-        assert model_provider_usage_sources(flow) == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 20),
-            ("gpt-5.5", "tokens.output", 12),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-        source_entries = model_usage_source_entries(flow)
-        assert len(source_entries) == 2
-        first_entry = next(
-            entry for entry in source_entries if entry["usage"]["tokens.input"] == 20
-        )
-        repeated_entry = next(
-            entry for entry in source_entries if entry["usage"]["tokens.input"] == 10
-        )
-        assert first_entry["source_id"] == repeated_entry["source_id"]
-        assert first_entry["provider_response_id"] == "resp_ws_1"
-        assert {event["source_idempotency_key"] for event in first_entry["usage_events"]} == {
-            event["source_idempotency_key"] for event in repeated_entry["usage_events"]
-        }
-        assert all(event["buffer_accepted"] is True for event in first_entry["usage_events"])
-        assert all(event["buffer_accepted"] is False for event in repeated_entry["usage_events"])
-        first_observations = first_entry["model_usage_observations"]
-        repeated_observations = repeated_entry["model_usage_observations"]
-        assert {observation["source_idempotency_key"] for observation in first_observations} == {
-            observation["source_idempotency_key"] for observation in repeated_observations
-        }
-        assert all(observation["buffer_accepted"] is True for observation in first_observations)
-        assert all(observation["buffer_accepted"] is False for observation in repeated_observations)
-
-    def test_model_websocket_late_same_id_frame_can_add_new_category_after_release(
-        self, tmp_path, real_flow
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "id": "resp_ws_1",
-                        "model": "gpt-5.5",
-                        "usage": {"input_tokens": 20},
-                    },
-                }
-            ).encode(),
-            json.dumps(
-                {
-                    "type": "response.done",
-                    "response": {
-                        "id": "resp_ws_1",
-                        "model": "gpt-5.5",
-                        "usage": {"output_tokens": 12},
-                    },
-                }
-            ).encode(),
-        )
-
-        assert model_provider_usage_sources(flow) == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 20),
-            ("gpt-5.5", "tokens.output", 12),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-
-    def test_model_websocket_late_output_reuses_long_context_tier_for_duplicates(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "id": "resp_ws_long_context",
-                        "model": "gpt-5.5",
-                        "usage": {"input_tokens": 272_001},
-                    },
-                }
-            ).encode(),
-            json.dumps(
-                {
-                    "type": "response.done",
-                    "response": {
-                        "id": "resp_ws_long_context",
-                        "model": "gpt-5.5",
-                        "usage": {"output_tokens": 12},
-                    },
-                }
-            ).encode(),
-            json.dumps(
-                {
-                    "type": "response.done",
-                    "response": {
-                        "id": "resp_ws_long_context",
-                        "model": "gpt-5.5",
-                        "usage": {"output_tokens": 7},
-                    },
-                }
-            ).encode(),
-        )
-
-        expected_billing_rows = [
-            ("gpt-5.5", "tokens.input.long_context", 272_001),
-            ("gpt-5.5", "tokens.output.long_context", 12),
-        ]
-        _assert_usage_event_rows(
-            webhook.usage_events(),
-            "provider",
-            expected_billing_rows,
-        )
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            [
-                ("gpt-5.5", "tokens.input", 272_001),
-                ("gpt-5.5", "tokens.output", 12),
-            ],
-        )
-        underbilling_entries = [
-            entry
-            for entry in read_jsonl_entries_after_flush(
-                Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-            )
-            if entry.get("type") == "usage_underbilling"
-        ]
-        assert underbilling_entries == []
-
-    def test_model_websocket_explicit_zero_input_selects_base_tier_for_late_output(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "id": "resp_ws_zero_input",
-                        "model": "gpt-5.5",
-                        "usage": {"input_tokens": 0},
-                    },
-                }
-            ).encode(),
-            json.dumps(
-                {
-                    "type": "response.done",
-                    "response": {
-                        "id": "resp_ws_zero_input",
-                        "model": "gpt-5.5",
-                        "usage": {"output_tokens": 12},
-                    },
-                }
-            ).encode(),
-        )
-
-        _assert_usage_event_rows(
-            webhook.usage_events(),
-            "provider",
-            [("gpt-5.5", "tokens.output", 12)],
-        )
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            [("gpt-5.5", "tokens.output", 12)],
-        )
-
-    def test_model_websocket_output_without_input_skips_unclassifiable_billing(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            json.dumps(
-                {
-                    "type": "response.done",
-                    "response": {
-                        "id": "resp_ws_missing_input",
-                        "model": "gpt-5.5",
-                        "usage": {"output_tokens": 12},
-                    },
-                }
-            ).encode(),
-        )
-
-        assert webhook.usage_events() == []
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            [("gpt-5.5", "tokens.output", 12)],
-        )
-        [entry] = [
-            entry
-            for entry in read_jsonl_entries_after_flush(
-                Path(flow.metadata[metadata_keys.VM_PROXY_LOG_PATH])
-            )
-            if entry.get("type") == "usage_underbilling"
-        ]
-        assert entry["reason"] == "model_long_context_tier_unresolved"
-        assert entry["underbilling_class"] == "risk"
-        assert entry["run_id"] == "run-abc-123"
-        assert entry["provider"] == "gpt-5.5"
-
-    def test_model_websocket_tier_state_is_bounded_and_terminally_released(
-        self,
-        tmp_path,
-        real_flow,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        for index in range(101):
-            feed_websocket_server_message(
-                flow,
-                json.dumps(
-                    {
-                        "type": "response.completed",
-                        "response": {
-                            "id": f"resp_ws_tier_{index}",
-                            "model": "gpt-5.5",
-                            "usage": {"input_tokens": 0},
-                        },
-                    }
-                ).encode(),
-            )
-
-        tiers = flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE_TIERS]
-        assert isinstance(tiers, dict)
-        assert len(tiers) == 100
-        assert "resp_ws_tier_0" not in tiers
-        assert tiers["resp_ws_tier_100"].tier == "base"
-        assert tiers["resp_ws_tier_100"].fast is False
-        assert tiers["resp_ws_tier_100"].committed is False
-
-        mitm_addon.websocket_end(flow)
-
-        assert metadata_keys.MODEL_PROVIDER_USAGE_TIERS not in flow.metadata
-
-    @pytest.mark.parametrize(
-        (
-            "input_tokens",
-            "cached_tokens",
-            "cache_write_tokens",
-            "service_tier",
-            "category_suffix",
-        ),
-        [
-            pytest.param(100, 20, 30, None, "", id="base"),
-            pytest.param(
-                300_000,
-                20_000,
-                30_000,
-                None,
-                ".long_context",
-                id="long-context",
-            ),
-            pytest.param(100, 20, 30, "priority", ".fast", id="fast"),
-            pytest.param(
-                300_000,
-                20_000,
-                30_000,
-                "priority",
-                ".long_context.fast",
-                id="long-context-fast",
-            ),
-        ],
-    )
-    def test_model_websocket_late_same_id_snapshot_does_not_mix_input_partition(
-        self,
-        tmp_path,
-        real_flow,
-        input_tokens,
-        cached_tokens,
-        cache_write_tokens,
-        service_tier,
-        category_suffix,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "gpt-5.6-sol"
-        mitm_addon.responseheaders(flow)
-        first_response = {
-            "id": "resp_ws_partition",
-            "model": "gpt-5.6-sol",
-            "usage": {
-                "input_tokens": input_tokens,
-                "output_tokens": 0,
-                "input_tokens_details": {"cached_tokens": cached_tokens},
-            },
-        }
-        if service_tier is not None:
-            first_response["service_tier"] = service_tier
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": first_response,
-                }
-            ).encode(),
-            json.dumps(
-                {
-                    "type": "response.done",
-                    "response": {
-                        "id": "resp_ws_partition",
-                        "model": "gpt-5.6-sol",
-                        "usage": {
-                            "input_tokens": input_tokens,
-                            "output_tokens": 40,
-                            "input_tokens_details": {
-                                "cached_tokens": cached_tokens,
-                                "cache_write_tokens": cache_write_tokens,
-                            },
-                        },
-                    },
-                }
-            ).encode(),
-        )
-
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert len(events) == len(by_category) == 3
-        assert len({event["idempotencyKey"] for event in events}) == 3
-        assert by_category == {
-            f"tokens.input{category_suffix}": input_tokens - cached_tokens,
-            f"tokens.output{category_suffix}": 40,
-            f"tokens.cache_read{category_suffix}": cached_tokens,
-        }
-        for event in events:
-            uuid.UUID(event["idempotencyKey"])
-        observation_events = webhook.model_usage_observation_events()
-        observations_by_category = compact_observation_quantities(observation_events)
-        assert len(observation_events) == 2
-        assert len({event["idempotencyKey"] for event in observation_events}) == 2
-        assert observations_by_category == {
-            "tokens.input": input_tokens - cached_tokens,
-            "tokens.output": 40,
-            "tokens.cache_read": cached_tokens,
-        }
-        for event in observation_events:
-            uuid.UUID(event["idempotencyKey"])
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-        assert model_provider_usage_sources(flow) == {}
-
-    def test_model_websocket_zero_frame_does_not_commit_base_tier(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "id": "resp_ws_1",
-                        "model": "gpt-5.5",
-                        "usage": {"input_tokens": 0, "output_tokens": 0},
-                    },
-                }
-            ).encode(),
-            json.dumps(
-                {
-                    "type": "response.done",
-                    "response": {
-                        "id": "resp_ws_1",
-                        "usage": {"input_tokens": 272_001, "output_tokens": 12},
-                    },
-                }
-            ).encode(),
-        )
-
-        assert model_provider_usage_sources(flow) == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input.long_context", 272_001),
-            ("gpt-5.5", "tokens.output.long_context", 12),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(
-            webhook.model_usage_observation_events(),
-            "model",
-            [
-                ("gpt-5.5", "tokens.input", 272_001),
-                ("gpt-5.5", "tokens.output", 12),
-            ],
-        )
-
-    def test_model_websocket_zero_frame_without_model_releases_source(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_server_message(
-                flow,
-                _openai_websocket_zero_usage_frame("resp_ws_zero_no_model", model=None),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        assert webhook.request_count == 0
-        assert model_provider_usage_sources(flow) == {}
-
-    def test_model_websocket_zero_frame_with_flow_model_releases_source(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        flow.metadata[metadata_keys.MODEL_USAGE_PROVIDER] = "gpt-flow-model"
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_server_message(
-                flow,
-                _openai_websocket_zero_usage_frame(
-                    "resp_ws_zero_flow_model", model="gpt-frame-model"
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        assert webhook.request_count == 0
-        assert model_provider_usage_sources(flow) == {}
-
-    def test_model_websocket_zero_frames_are_bounded(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            for index in range(70):
-                feed_websocket_server_message(
-                    flow,
-                    _openai_websocket_zero_usage_frame(
-                        f"resp_ws_zero_{index}", model=f"gpt-zero-{index}"
-                    ),
-                )
-            usage.flush_usage_events(trigger="test")
-
-        assert webhook.request_count == 0
-        usage_sources = model_provider_usage_sources(flow)
-        assert usage_sources == {}
-
-    def test_model_websocket_flow_model_reports_later_positive_usage_without_zero_hint(
-        self, tmp_path, real_flow
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            for index in range(70):
-                feed_websocket_server_message(
-                    flow,
-                    _openai_websocket_zero_usage_frame(
-                        f"resp_ws_zero_{index}", model=f"gpt-zero-{index}"
-                    ),
-                )
-
-            assert "resp_ws_zero_0" not in model_provider_usage_sources(flow)
-
-            feed_websocket_server_message(
-                flow,
-                json.dumps(
-                    {
-                        "type": "response.done",
-                        "response": {
-                            "id": "resp_ws_zero_0",
-                            "usage": {"input_tokens": 20, "output_tokens": 12},
-                        },
-                    }
-                ).encode(),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        usage_sources = model_provider_usage_sources(flow)
-        assert "resp_ws_zero_0" not in usage_sources
-        assert usage_sources == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 20),
-            ("gpt-5.5", "tokens.output", 12),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-
-    def test_model_websocket_text_frame_reports_usage(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_server_text_message(
-                flow,
-                json.dumps(
-                    {
-                        "type": "response.completed",
-                        "response": {
-                            "id": "resp_ws_text",
-                            "model": "gpt-5.5",
-                            "usage": {"input_tokens": 3, "output_tokens": 2},
-                        },
-                    }
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        assert model_provider_usage_sources(flow) == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 3),
-            ("gpt-5.5", "tokens.output", 2),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-
-    def test_model_websocket_valid_frame_replaces_invalid_usage_sources_metadata(
-        self, tmp_path, real_flow
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE_SOURCES] = "invalid"
-
-        with self._usage_webhook_api() as webhook:
-            feed_websocket_server_message(
-                flow,
-                openai_websocket_usage_frame(
-                    "resp_ws_invalid_sources",
-                    input_tokens=10,
-                    output_tokens=4,
-                ),
-            )
-            usage.flush_usage_events(trigger="test")
-
-        assert model_provider_usage_sources(flow) == {}
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 10),
-            ("gpt-5.5", "tokens.output", 4),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-
-    def test_model_websocket_same_id_zero_after_positive_does_not_double_bill(
-        self, tmp_path, real_flow
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            openai_websocket_usage_frame(
-                "resp_ws_1",
-                input_tokens=20,
-                output_tokens=12,
-            ),
-            openai_websocket_usage_frame(
-                "resp_ws_1",
-                input_tokens=0,
-                output_tokens=0,
-            ),
-        )
-
-        assert model_provider_usage_sources(flow) == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 20),
-            ("gpt-5.5", "tokens.output", 12),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-
-    def test_full_pipeline_model_websocket_uses_context_model_for_response_ids(
-        self, tmp_path, real_flow
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            openai_websocket_usage_frame(
-                "resp_ws_1",
-                input_tokens=10,
-                output_tokens=4,
-                model="gpt-5.5",
-            ),
-            openai_websocket_usage_frame(
-                "resp_ws_2",
-                input_tokens=3,
-                output_tokens=2,
-                model="gpt-5.6-luna",
-            ),
-        )
-
-        assert model_provider_usage_sources(flow) == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 10),
-            ("gpt-5.5", "tokens.output", 4),
-            ("gpt-5.5", "tokens.input", 3),
-            ("gpt-5.5", "tokens.output", 2),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-
-    def test_full_pipeline_model_websocket_reports_id_and_missing_id_usage(
-        self, tmp_path, real_flow
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "model": "gpt-5.5",
-                        "usage": {"input_tokens": 7, "output_tokens": 1},
-                    },
-                }
-            ).encode(),
-            openai_websocket_usage_frame(
-                "resp_ws_1",
-                input_tokens=10,
-                output_tokens=4,
-                model="gpt-5.5",
-            ),
-        )
-
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {
-            "model": "gpt-5.5",
-            "tokens.input": 7,
-            "tokens.output": 1,
-        }
-        assert model_provider_usage_sources(flow) == {}
-        expected_rows = [
-            ("gpt-5.5", "tokens.input", 7),
-            ("gpt-5.5", "tokens.output", 1),
-            ("gpt-5.5", "tokens.input", 10),
-            ("gpt-5.5", "tokens.output", 4),
-        ]
-        _assert_usage_event_rows(webhook.usage_events(), "provider", expected_rows)
-        _assert_usage_event_rows(webhook.model_usage_observation_events(), "model", expected_rows)
-        source_entries = model_usage_source_entries(flow)
-        assert len(source_entries) == 2
-        [source_preserving_entry] = [
-            entry for entry in source_entries if entry["buffer_mode"] == "source"
-        ]
-        [aggregate_entry] = [
-            entry for entry in source_entries if entry["buffer_mode"] == "aggregate"
-        ]
-        assert source_preserving_entry["provider_response_id"] == "resp_ws_1"
-        assert aggregate_entry["provider_response_id"] is None
-        assert aggregate_entry["source_id"] == flow.id
-        assert aggregate_entry["transport"] == "websocket"
-        webhook_usage_keys = {event["idempotencyKey"] for event in webhook.usage_events()}
-        assert {
-            event["source_idempotency_key"] for event in source_preserving_entry["usage_events"]
-        }.issubset(webhook_usage_keys)
-        assert {
-            event["source_idempotency_key"] for event in aggregate_entry["usage_events"]
-        }.isdisjoint(webhook_usage_keys)
-        assert all(
-            event["buffer_accepted"] is True
-            for entry in source_entries
-            for event in entry["usage_events"]
-        )
-
-    @pytest.mark.parametrize(
-        "later_cache_write_tokens",
-        [0, None],
-        ids=["zero", "omitted"],
-    )
-    def test_model_websocket_missing_id_raw_snapshots_preserve_input_partition(
-        self,
-        tmp_path,
-        real_flow,
-        later_cache_write_tokens,
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        later_input_details = {"cached_tokens": 20}
-        if later_cache_write_tokens is not None:
-            later_input_details["cache_write_tokens"] = later_cache_write_tokens
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "model": "gpt-5.5",
-                        "usage": {
-                            "input_tokens": 100,
-                            "output_tokens": 0,
-                            "input_tokens_details": {
-                                "cached_tokens": 20,
-                                "cache_write_tokens": 30,
-                            },
-                        },
-                    },
-                }
-            ).encode(),
-            json.dumps(
-                {
-                    "type": "response.done",
-                    "response": {
-                        "model": "gpt-5.5",
-                        "usage": {
-                            "input_tokens": 100,
-                            "output_tokens": 40,
-                            "input_tokens_details": later_input_details,
-                        },
-                    },
-                }
-            ).encode(),
-        )
-
-        events = webhook.usage_events()
-        by_category = {event["category"]: event["quantity"] for event in events}
-        assert len(events) == len(by_category) == 4
-        assert by_category == {
-            "tokens.input": 50,
-            "tokens.output": 40,
-            "tokens.cache_read": 20,
-            "tokens.cache_creation": 30,
-        }
-        assert (
-            by_category["tokens.input"]
-            + by_category["tokens.cache_read"]
-            + by_category["tokens.cache_creation"]
-            == 100
-        )
-        observation_events = webhook.model_usage_observation_events()
-        observations_by_category = compact_observation_quantities(observation_events)
-        assert len(observation_events) == 1
-        assert observations_by_category == by_category
-        assert model_provider_usage_sources(flow) == {}
-
-    def test_full_pipeline_model_websocket_zero_frame_preserves_billed_usage_and_id(
-        self, tmp_path, real_flow
-    ):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-
-        webhook = self._run_websocket_messages_and_end(
-            flow,
-            json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "id": "resp_ws_1",
-                        "model": "gpt-5.5",
-                        "usage": {
-                            "input_tokens": 100,
-                            "output_tokens": 40,
-                        },
-                    },
-                }
-            ).encode(),
-            json.dumps(
-                {
-                    "type": "response.done",
-                    "response": {
-                        "id": "resp_ws_empty",
-                        "model": "gpt-5.6-luna",
-                        "usage": {
-                            "input_tokens": 0,
-                            "output_tokens": 0,
-                            "input_tokens_details": {"cached_tokens": 0},
-                        },
-                    },
-                }
-            ).encode(),
-        )
-
-        assert model_provider_usage_sources(flow) == {}
-        _assert_usage_event_rows(
-            webhook.usage_events(),
-            "provider",
-            [
-                ("gpt-5.5", "tokens.input", 100),
-                ("gpt-5.5", "tokens.output", 40),
-            ],
-        )
-
-    def test_model_websocket_ignores_client_messages(self, tmp_path, real_flow):
-        flow = make_openai_responses_websocket_flow(real_flow, tmp_path)
-        mitm_addon.responseheaders(flow)
-        set_websocket_message(
-            flow,
-            from_client=True,
-            content=json.dumps(
-                {
-                    "type": "response.completed",
-                    "response": {
-                        "id": "resp_ws_1",
-                        "model": "gpt-5.5",
-                        "usage": {"input_tokens": 50, "output_tokens": 20},
-                    },
-                }
-            ).encode(),
-        )
-
-        webhook = self._run_websocket_message_and_end(flow)
-
-        assert webhook.request_count == 0
-        assert flow.metadata[metadata_keys.MODEL_PROVIDER_USAGE] == {}
-        assert model_provider_usage_sources(flow) == {}
